week_5/zoomcamp/pipeline/assets/ingestion/requirements.txt
requests
pandas
pyarrow
python-dateutil